requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
mysql-connector-python==8.3.0
requests==2.31.0
urllib3==2.2.0
